Split manifest module into versioning submodules
Move src/manifest/ → src/versioning/ and break the ~1100-line mod.rs into
focused submodules so each concept reads as a self-contained unit with its
own tests:

  versioning/mod.rs       shared types (FileId, Level, VersionEdit,
                          SstableMetadata, sst_path, NUM_LEVELS) plus
                          re-exports of the submodule public items
  versioning/guard.rs     SstFileGuard refcounted file handle
  versioning/state.rs     VersionState and SharedVersion (CoW wrapper)
  versioning/manifest.rs  Manifest log, ManifestError, replay,
                          serialize_edit, sweep_orphans, EditType

Tests move with their type — guard, state, and manifest each get a tests
submodule with its own tmp_dir prefix. Callers in engine.rs, flusher,
compaction, and main.rs are updated to import from crate::versioning.
No behaviour change.
