+++ src/compaction/mod.rs
@@ -9,7 +9,7 @@ use std::sync::mpsc::Receiver;
 
 use crate::core::{InternalKey, KvIterator, Record};
 use crate::engine::EngineCore;
-use crate::manifest::{SstableMetadata, VersionState, sst_path};
+use crate::versioning::{SstableMetadata, VersionState, sst_path};
 use crate::sstable::block::Block;
 use crate::sstable::{FOOTER_SIZE, INDEX_OFFSET_SIZE, MAGIC_NUMBER, META_OFFSET_SIZE};
 use crate::sstable::writer::SsTableBuilder;
